requirements.txt
flask
requests
firebase-admin
flask-cors
boto3
python-dotenv
